fixed test + formatting

+++ src/peripheral/BaseBalancerCompounder.sol
@@ -25,25 +25,14 @@ abstract contract BaseBalancerCompounder {
 
         uint256 amount;
         uint256 rewLen = sellPaths.length;
-        for (uint256 i = 0; i < rewLen; ) {
-            amount = IERC20(address(sellPaths[i].assets[0])).balanceOf(
-                address(this)
-            );
+        for (uint256 i = 0; i < rewLen;) {
+            amount = IERC20(address(sellPaths[i].assets[0])).balanceOf(address(this));
 
             if (amount > 0) {
                 // Decode since nested struct[] isnt allowed in storage
-                BatchSwapStep[] memory swaps = abi.decode(
-                    sellPaths[i].swaps,
-                    (BatchSwapStep[])
-                );
+                BatchSwapStep[] memory swaps = abi.decode(sellPaths[i].swaps, (BatchSwapStep[]));
 
-                BalancerTradeLibrary.trade(
-                    router,
-                    swaps,
-                    sellPaths[i].assets,
-                    sellPaths[i].limits,
-                    amount
-                );
+                BalancerTradeLibrary.trade(router, swaps, sellPaths[i].assets, sellPaths[i].limits, amount);
             }
 
             unchecked {
@@ -52,10 +41,7 @@ abstract contract BaseBalancerCompounder {
         }
     }
 
-    function setBalancerTradeValues(
-        address newBalancerVault,
-        TradePath[] memory newTradePaths
-    ) internal {
+    function setBalancerTradeValues(address newBalancerVault, TradePath[] memory newTradePaths) internal {
         // Remove old rewardToken allowance
         uint256 rewardTokenLen = _rewardTokens.length;
         if (rewardTokenLen > 0) {
@@ -64,7 +50,7 @@ abstract contract BaseBalancerCompounder {
             address[] memory oldRewardTokens = _rewardTokens;
 
             // void approvals
-            for (uint256 i = 0; i < rewardTokenLen; ) {
+            for (uint256 i = 0; i < rewardTokenLen;) {
                 IERC20(oldRewardTokens[i]).approve(oldBalancerVault, 0);
 
                 unchecked {
@@ -80,7 +66,7 @@ abstract contract BaseBalancerCompounder {
         // Add new allowance + state
         address newRewardToken;
         rewardTokenLen = newTradePaths.length;
-        for (uint i; i < rewardTokenLen; ) {
+        for (uint256 i; i < rewardTokenLen;) {
             newRewardToken = address(newTradePaths[i].assets[0]);
 
             IERC20(newRewardToken).approve(newBalancerVault, type(uint256).max);
